requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-whois==0.8.0
pydantic==2.5.0
